+++ spring_boot_backend/Servers/Pivotal tc Server Developer Edition v4.0-config/jmxremote.access
@@ -1,0 +1,2 @@
+#admin readonly
+admin readwrite
