+++ spring_boot_backend/Servers/Pivotal tc Server Developer Edition v4.0-config/jmxremote.password
@@ -1,0 +1,2 @@
+# For security reasons the "admin" role has a random password unless the property is overridden at instance creation time.
+admin avgxqvffmjkmdggkrrfiiibmdmgsqujhygiujjgnisrfp
