+++ spring_boot_backend/Servers/Pivotal tc Server Developer Edition v4.0-config/tomcat-users.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tomcat-users/>
