+++ spring_boot_backend/src/main/resources/application.properties
@@ -1,22 +1,23 @@
 #Tomcat server port number 8080
 #server.port=7070
 #context path default value : empty
-#server.servlet.context-path=/ems
+#server.servlet.context-path=/test_boot
 server.servlet.session.persistent=false
 #DB properties
-spring.datasource.url=jdbc:mysql://localhost:3306/advjava2?createDatabaseIfNotExist=true&useSSL=false&allowPublicKeyRetrieval=true
+spring.datasource.url=jdbc:mysql://localhost:3306/ems_project?createDatabaseIfNotExist=true&useSSL=false&allowPublicKeyRetrieval=true
 spring.datasource.username=root
-spring.datasource.password=root
+spring.datasource.password=manager
 # JPA properties
 spring.jpa.show-sql = true
 spring.jpa.hibernate.ddl-auto = update
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL8Dialect
 #logging.level.org.springframework.orm.hibernate5=DEBUG
-
-# Spring MVC ViewResolver related
-#spring.mvc.view.prefix=/WEB-INF/views
-#spring.mvc.view.suffix=.jsp
 #Disable open session in view 
 spring.jpa.open-in-view=false
 #property to disable hibernateLazyInitializer during initialized proxies
 #spring.jackson.serialization.FAIL_ON_EMPTY_BEANS=false
+#image uploading
+spring.servlet.multipart.max-file-size=10MB
+spring.servlet.multipart.max-request-size=11MB
+#file upload location : custom property
+file.upload.location=images/
