--- spring_boot_backend_template/src/main/java/com/app/Application.java
@@ -1,23 +1,0 @@
-package com.app;
-
-import org.modelmapper.ModelMapper;
-import org.modelmapper.convention.MatchingStrategies;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-@SpringBootApplication
-public class Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-
-	@Bean // equivalent to <bean id ..../> in xml file
-	public ModelMapper mapper() {
-		ModelMapper modelMapper = new ModelMapper();		modelMapper.getConfiguration().setMatchingStrategy(MatchingStrategies.STRICT)
-	.setPropertyCondition(Conditions.isNotNull());
-		return modelMapper;
-	}
-
-}
