--- spring_boot_backend_template/src/test/java/com/app/ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
